requirements.txt
a2a-sdk>=0.2.4
pydantic>=2.0.0
uvicorn>=0.17.0
starlette>=0.27.0
python-dotenv>=1.0.0
httpx>=0.24.0
process-gpt-agent-sdk==0.3.4
process-gpt-llm-factory>=1.0.0
pytest>=8.4.2
supabase>=2.0.0